# contract/openapi.yaml
openapi: 3.0.3
info:
  title: Ratatoskr API
  version: 2.1.0
  description: |
    Ratatoskr bridges an Audiobookshelf (ABS) server and Sonos speakers. This
    document describes the HTTP API that Ratatoskr's own clients (the Android app,
    and any future client) use to browse the library, pick a speaker, and control
    playback. Audio itself is streamed directly from Audiobookshelf to the Sonos
    speakers and does not pass through this API.

    This contract is the single source of truth for client/server communication.
    Server and clients are versioned and released independently, and must remain
    compatible across a range of versions. Therefore:

      - The API is versioned in the URL path, kept in one place (servers.url).
        Breaking changes go into a new major version and a new path, which may be served
        alongside the old one during a transition window.
      - Backwards-compatible additions (new optional fields, new endpoints) are minor
        version bumps within the current major. Clients must ignore unknown fields.
      - A CI check (oasdiff) fails any change that breaks compatibility without a
        major version bump.

    This document is the /v2 surface: Ratatoskr-native sessions, where the server is the
    sole holder of Audiobookshelf tokens and the client holds one opaque Ratatoskr token
    (ADR-0001). Its predecessor /v1 - a shared Audiobookshelf token pair with a
...
    tag, not by this file.
  license:
    name: GPL-3.0-or-later
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /auth/device-sessions:
    get:
      operationId: listDeviceSessions
      summary: List the caller's own device sessions
      description: |
        The signed-in user's own device sessions - every device that is currently signed in as
        this user, the caller's included. A hygiene overview, for spotting a device that was
        reset or retired but still holds a session; it is not a threat-detection tool. A device
        session is one sign-in, not a physical device: a phone that was reset and signed in
        again appears twice. Another user's sessions are never listed.

        Ordered most recently signed in first.
      tags: [system]
      responses:
        "200":
          description: The caller's device sessions.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeviceSessionList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/device-sessions/{id}:
    delete:
      operationId: endDeviceSession
      summary: End one of the caller's own device sessions
      description: |
        Has exactly the effect sign-out has on that device: its token is invalid immediately,
        and the Audiobookshelf session Ratatoskr holds for the user is released only when this
        was the user's last device. The caller may end its own entry this way too, with the
        same effect as `POST /auth/logout`.

        Unlike `POST /auth/logout` this is **not** idempotent: an id that names no device
        session answers 404. So does the id of another user's device session, in the same
        way, so the existence of foreign ids is not revealed.
      tags: [system]
      parameters:
        - $ref: "#/components/parameters/DeviceSessionId"
      responses:
        "204":
          description: The device session was ended.
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        The caller's Ratatoskr token, obtained from POST /auth/login. It is opaque, does not
        expire and is never rotated: a client stores it once and sends it unchanged until
        POST /auth/logout. Ratatoskr resolves it to the Audiobookshelf session it holds for
        this device, so the library view and progress stay scoped to the user who signed in.
        Required on every endpoint except /health, /auth/login, and GET /speakers.
  parameters:
    DeviceSessionId:
      name: id
      in: path
      required: true
      description: Device session id, as listed by `GET /auth/device-sessions`.
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: The request was malformed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: |
        The credential was rejected. On the bearer-protected endpoints that is the bearer token; on
        `login` (which takes no bearer) it is the Audiobookshelf username and password. The body's
        `code` separates the cases a client must handle differently — see UnauthorizedError.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/UnauthorizedError" }
    NotFound:
      description: The resource was not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    DeviceSession:
      type: object
      required: [id, createdAt, current]
      description: |
        One sign-in of the caller's user, as listed by `GET /auth/device-sessions`. Unrelated to
        the playback Session: this is about who is signed in where, that is about what is playing.
      properties:
        id:
          type: string
          format: uuid
          description: |
            Public identifier of this device session, for `DELETE /auth/device-sessions/{id}`.
            Unrelated to the bearer token, which it does not reveal.
        deviceName:
          type: string
          description: |
            The label the device sent at sign-in. Absent when it sent none, and for sign-ins that
            predate this field; a client shows "unknown device" and `createdAt` then.
        createdAt:
          type: string
          format: date-time
          description: When this device signed in.
        lastUsedAt:
          type: string
          format: date-time
          description: |
            When this device last made a request that proved its bearer. Only the device's own
            requests count. Absent when it has not been used since sign-in was recorded, and for
            sign-ins that predate this field.
        current:
          type: boolean
          description: True for the caller's own device session.
    DeviceSessionList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/DeviceSession" }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable, machine-readable error code.
        message:
          type: string
          description: Human-readable explanation.
    UnauthorizedError:
      # The 401 body, narrowed to the codes a client branches on: the two failures look alike
      # on the wire but call for opposite reactions, so the distinction has to be machine-
      # readable rather than left to the message (server SPEC section 8).
      # UPSTREAM_SESSION_LOST is spelled as ADR-0001 and SPEC section 8 spell it; the
      # lowercase codes predate that convention and stay as they are within this major.
      allOf:
        - $ref: "#/components/schemas/Error"
        - type: object
          properties:
            code:
              type: string
              enum: [unauthorized, UPSTREAM_SESSION_LOST]
              description: |
                `unauthorized` — no token, or one this server does not know (it was never
                issued, or sign-out revoked it); at `login`, the Audiobookshelf credentials
                were refused. The device is signed out, and the client asks for credentials as
                a fresh sign-in.

                `UPSTREAM_SESSION_LOST` — the token is valid, but the Audiobookshelf session
                Ratatoskr held for it has died (server↔Audiobookshelf contact lost for longer
                than Audiobookshelf's refresh window, or the username changed). Only the
                password can restore it: the client shows a targeted re-authentication prompt
                and keeps the rest of its state. Re-authenticating yields a new token.